fix: Move BitmapWithSize model to its unique file & remove unused model
