+++ Modules/LuxFilter/LuxFilter/Interfaces/IPipelineService.cs
@@ -1,5 +1,5 @@
 ﻿using LuxFilter.Algorithms.Interfaces;
-using SkiaSharp;
+using LuxFilter.Models;
 
 namespace LuxFilter.Interfaces
 {
@@ -19,19 +19,5 @@ namespace LuxFilter.Interfaces
         /// Compute scores for a collection of BitmapWithSize objects
         /// </summary>
         Task<List<double>> Compute(IEnumerable<BitmapWithSize> bitmapsWithSizes);
-    }
-
-    public class BitmapWithSize
-    {
-        public SKBitmap Bitmap { get; set; }
-        public int Height { get; set; }
-        public int Width { get; set; }
-
-        public BitmapWithSize(SKBitmap bitmap)
-        {
-            Bitmap = bitmap;
-            Height = bitmap.Height;
-            Width = bitmap.Width;
-        }
     }
 }
